flow-types to type.js file
