+++ .eslintrc
@@ -8,7 +8,8 @@
       {
         "extensions": [".js", ".jsx"]
       }
-    ]
+    ],
+    "react/forbid-prop-types": [0, { "forbid": ["object"] }]
   },
   "env": {
     "es6": true,
