+++ src/Textarea.jsx
@@ -6,58 +6,24 @@ import getCaretCoordinates from 'textarea-caret';
 
 import Listeners, { KEY_CODES } from './listener';
 import List from './List';
+import type {
+  TextareaProps,
+  TextareaState,
+  getTextToReplaceType,
+  triggerType,
+  settingType,
+} from './types';
 
-type dataProviderType = string =>
-  | Promise<Array<Object | string>>
-  | Array<Object | string>;
-
-type settingType = {
-  component: React$StatelessFunctionalComponent<*>,
-  dataProvider: dataProviderType,
-  output?: (Object | string, ?string) => string,
-};
-
-type getTextToReplaceType = (Object | string) => string;
-
-type triggerType = {
-  [string]: {|
-    output?: (Object | string, ?string) => string,
-    dataProvider: dataProviderType,
-    component: React$StatelessFunctionalComponent<*>,
-  |},
-};
-
-type Props = {
-  trigger: triggerType,
-  loadingComponent: React$StatelessFunctionalComponent<*>,
-  onChange: ?(SyntheticEvent<*> | Event) => void,
-  minChar: ?number,
-  value?: string,
-  style: ?Object,
-  containerStyle: ?Object,
-  className: ?string,
-};
-
-type State = {
-  currentTrigger: ?string,
-  top: number,
-  left: number,
-  actualToken: string,
-  data: ?Array<Object | string>,
-  value: string,
-  dataLoading: boolean,
-  selectionEnd: number,
-  selectionStart: number,
-  component: ?React$StatelessFunctionalComponent<*>,
-};
-
-class ReactTextareaAutocomplete extends React.Component<Props, State> {
+class ReactTextareaAutocomplete extends React.Component<TextareaProps, TextareaState> {
   static defaultProps = {
-    value: '',
+    containerStyle: {},
     minChar: 1,
+    onChange: null,
+    style: {},
+    value: '',
   };
 
-  constructor(props: Props) {
+  constructor(props: TextareaProps) {
     super(props);
 
     Listeners.add(KEY_CODES.ESC, () => this.closeAutocomplete());
@@ -76,7 +42,7 @@ class ReactTextareaAutocomplete extends React.Component<Props, State> {
     }
   }
 
-  state: State = {
+  state: TextareaState = {
     top: 0,
     left: 0,
     currentTrigger: null,
@@ -93,7 +59,7 @@ class ReactTextareaAutocomplete extends React.Component<Props, State> {
     Listeners.startListen();
   }
 
-  componentWillReceiveProps(nextProps: Props) {
+  componentWillReceiveProps(nextProps: TextareaProps) {
     this.update(nextProps);
   }
 
@@ -234,7 +200,7 @@ class ReactTextareaAutocomplete extends React.Component<Props, State> {
     return data;
   };
 
-  update({ value, trigger }: Props) {
+  update({ value, trigger }: TextareaProps) {
     const { value: oldValue } = this.state;
     const { trigger: oldTrigger } = this.props;
 
@@ -324,7 +290,7 @@ class ReactTextareaAutocomplete extends React.Component<Props, State> {
     );
   };
 
-  props: Props;
+  props: TextareaProps;
 
   textareaRef: HTMLInputElement;
 
@@ -410,8 +376,12 @@ const triggerPropsCheck = ({ trigger }: { trigger: triggerType }) => {
 };
 
 ReactTextareaAutocomplete.propTypes = {
-  trigger: triggerPropsCheck, //eslint-disable-line
+  containerStyle: PropTypes.object,
   loadingComponent: PropTypes.func.isRequired,
+  minChar: PropTypes.number,
+  onChange: PropTypes.func,
+  style: PropTypes.object,
+  trigger: triggerPropsCheck, //eslint-disable-line
   value: PropTypes.string,
 };
